team6/requirements.txt
Django==2.1.8
django-polymorphic==2.1.2
pytz==2019.3
